backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
beautifulsoup4==4.12.2
spacy==3.7.2
PyPDF2==3.0.1
openai>=1.50.0
requests==2.31.0
python-dotenv==1.0.0
torch>=2.0.0
transformers>=4.30.0
scikit-learn>=1.3.0
numpy>=1.24.0

